Split the demo app into two (`demo.eda`, `demo.search`)
