+++ src/demo/apps/eda/📊_Product_Catalogue.py
