+++ src/demo/apps/search/pages/2_🤖_Synonyms.py
