+++ src/demo/apps/search/pages/7_🔍_Search.py
