+++ tasks/__init__.py
@@ -1,6 +1,6 @@
 from invoke import Collection, task
 
-from tasks import data_tasks, es_tasks, gcloud_tasks, synonyms_tasks, vespa_tasks
+from tasks import data_tasks, demo_tasks, es_tasks, gcloud_tasks, synonyms_tasks, vespa_tasks
 
 
 @task
@@ -31,16 +31,11 @@ def lint(c):
     print("Done")
 
 
-@task
-def demo(c):
-    c.run("poetry run streamlit run src/demo/🏠_Home.py")
-
-
 ns = Collection()
 ns.add_task(format)
 ns.add_task(lint)
-ns.add_task(demo)
 ns.add_collection(Collection.from_module(data_tasks, name="data"))
+ns.add_collection(Collection.from_module(demo_tasks, name="demo"))
 ns.add_collection(Collection.from_module(es_tasks, name="es"))
 ns.add_collection(Collection.from_module(gcloud_tasks, name="gcloud"))
 ns.add_collection(Collection.from_module(synonyms_tasks, name="synonyms"))
